+++ database/install/myems_energy_plan_db.sql
@@ -1,31 +1,17 @@
+-- MyEMS Energy Plan Database
+
 -- ---------------------------------------------------------------------------------------------------------------------
--- WARNING: BACKUP YOUR DATABASE BEFORE UPGRADING
--- THIS SCRIPT IS ONLY FOR UPGRADING 4.3.0 TO 4.4.0
--- THE CURRENT VERSION CAN BE FOUND AT `myems_system_db`.`tbl_versions`
+-- Schema myems_energy_plan_db
 -- ---------------------------------------------------------------------------------------------------------------------
-
-START TRANSACTION;
-
-UPDATE myems_system_db.tbl_menus SET route='/energystoragepowerstation/details' WHERE id=1401;
-
-DELETE FROM myems_system_db.tbl_menus WHERE id=1401;
-INSERT INTO myems_system_db.tbl_menus (id, name, route, parent_menu_id, is_hidden)
-VALUES
-(1401, 'Energy Storage Power Station List','/energystoragepowerstation/list', 1400, 1),
-(1402,'Energy Storage Power Station Details','/energystoragepowerstation/details',1400,1),
-(1403, 'Energy Storage Power Station Reporting','/energystoragepowerstation/reporting', 1400, 1),
-(1404, 'Energy Storage Power Station Alarm','/energystoragepowerstation/alarm', 1400, 1),
-(1405, 'Energy Storage Power Station Maintenance','/energystoragepowerstation/maintenance', 1400, 1);
-
--- add columns
-ALTER TABLE `myems_fdd_db`.`tbl_web_messages` ADD `update_datetime_utc` DATETIME NULL AFTER `created_datetime_utc`;
-ALTER TABLE `myems_fdd_db`.`tbl_web_messages` ADD `end_datetime_utc` DATETIME NULL AFTER `created_datetime_utc`;
-ALTER TABLE `myems_fdd_db`.`tbl_web_messages` ADD `start_datetime_utc` DATETIME NULL AFTER `created_datetime_utc`;
-ALTER TABLE `myems_system_db`.`tbl_energy_storage_power_stations` ADD `is_cost_data_displayed` BOOL NOT NULL AFTER `svg`;
-ALTER TABLE `myems_system_db`.`tbl_microgrids` ADD `is_cost_data_displayed` BOOL NOT NULL AFTER `svg`;
-
--- Create Energy Plan Database
+DROP DATABASE IF EXISTS `myems_energy_plan_db` ;
 CREATE DATABASE IF NOT EXISTS `myems_energy_plan_db` CHARACTER SET 'utf8mb4' COLLATE 'utf8mb4_unicode_ci' ;
+USE `myems_energy_plan_db` ;
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_combined_equipment_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_input_category_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `combined_equipment_id` BIGINT NOT NULL,
@@ -36,6 +22,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_input_
 CREATE INDEX `tbl_combined_equipment_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_combined_equipment_input_category_hourly`
  (`combined_equipment_id`, `energy_category_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_combined_equipment_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_input_item_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -48,6 +39,11 @@ CREATE INDEX `tbl_combined_equipment_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_combined_equipment_input_item_hourly`
  (`combined_equipment_id`, `energy_item_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_combined_equipment_output_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_output_category_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_output_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `combined_equipment_id` BIGINT NOT NULL,
@@ -58,6 +54,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_combined_equipment_output
 CREATE INDEX `tbl_combined_equipment_output_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_combined_equipment_output_category_hourly`
  (`combined_equipment_id`, `energy_category_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_equipment_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_equipment_input_category_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_equipment_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -70,6 +71,11 @@ CREATE INDEX `tbl_equipment_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_equipment_input_category_hourly`
  (`equipment_id`, `energy_category_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_equipment_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_equipment_input_item_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_equipment_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `equipment_id` BIGINT NOT NULL,
@@ -80,6 +86,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_equipment_input_item_hour
 CREATE INDEX `tbl_equipment_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_equipment_input_item_hourly`
  (`equipment_id`, `energy_item_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_equipment_output_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_equipment_output_category_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_equipment_output_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -92,6 +103,11 @@ CREATE INDEX `tbl_equipment_output_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_equipment_output_category_hourly`
  (`equipment_id`, `energy_category_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_meter_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_meter_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_meter_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `meter_id` BIGINT NOT NULL,
@@ -102,6 +118,11 @@ CREATE INDEX `tbl_meter_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_meter_hourly`
  (`meter_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_offline_meter_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_offline_meter_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_offline_meter_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `offline_meter_id` BIGINT NOT NULL,
@@ -110,6 +131,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_offline_meter_hourly` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_offline_meter_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_offline_meter_hourly` (`offline_meter_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_shopfloor_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_shopfloor_input_category_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_shopfloor_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -122,6 +148,11 @@ CREATE INDEX `tbl_shopfloor_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_shopfloor_input_category_hourly`
  (`shopfloor_id`, `energy_category_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_shopfloor_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_shopfloor_input_item_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_shopfloor_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `shopfloor_id` BIGINT NOT NULL,
@@ -132,6 +163,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_shopfloor_input_item_hour
 CREATE INDEX `tbl_shopfloor_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_shopfloor_input_item_hourly`
  (`shopfloor_id`, `energy_item_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_space_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_space_input_category_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_space_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -144,6 +180,11 @@ CREATE INDEX `tbl_space_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_space_input_category_hourly`
  (`space_id`, `energy_category_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_space_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_space_input_item_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_space_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `space_id` BIGINT NOT NULL,
@@ -154,6 +195,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_space_input_item_hourly` 
 CREATE INDEX `tbl_space_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_space_input_item_hourly`
  (`space_id`, `energy_item_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_space_output_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_space_output_category_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_space_output_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -166,6 +212,11 @@ CREATE INDEX `tbl_space_output_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_space_output_category_hourly`
  (`space_id`, `energy_category_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_store_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_store_input_category_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_store_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `store_id` BIGINT NOT NULL,
@@ -176,6 +227,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_store_input_category_hour
 CREATE INDEX `tbl_store_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_store_input_category_hourly`
  (`store_id`, `energy_category_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_store_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_store_input_item_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_store_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -188,6 +244,11 @@ CREATE INDEX `tbl_store_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_store_input_item_hourly`
  (`store_id`, `energy_item_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_tenant_input_category_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_tenant_input_category_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_tenant_input_category_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `tenant_id` BIGINT NOT NULL,
@@ -198,6 +259,11 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_tenant_input_category_hou
 CREATE INDEX `tbl_tenant_input_category_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_tenant_input_category_hourly`
  (`tenant_id`, `energy_category_id`, `start_datetime_utc`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_tenant_input_item_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_tenant_input_item_hourly` ;
 
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_tenant_input_item_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
@@ -210,6 +276,11 @@ CREATE INDEX `tbl_tenant_input_item_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_tenant_input_item_hourly`
  (`tenant_id`, `energy_item_id`, `start_datetime_utc`);
 
+-- ---------------------------------------------------------------------------------------------------------------------
+-- Table `myems_energy_plan_db`.`tbl_virtual_meter_hourly`
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_energy_plan_db`.`tbl_virtual_meter_hourly` ;
+
 CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_virtual_meter_hourly` (
   `id` BIGINT NOT NULL AUTO_INCREMENT,
   `virtual_meter_id` BIGINT NOT NULL,
@@ -218,8 +289,3 @@ CREATE TABLE IF NOT EXISTS `myems_energy_plan_db`.`tbl_virtual_meter_hourly` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_virtual_meter_hourly_index_1`
  ON `myems_energy_plan_db`.`tbl_virtual_meter_hourly` (`virtual_meter_id`, `start_datetime_utc`);
-
--- UPDATE VERSION NUMBER
-UPDATE `myems_system_db`.`tbl_versions` SET version='4.4.0RC', release_date='2024-04-10' WHERE id=1;
-
-COMMIT;